production/requirements.txt
# Core AI and Agents
openai>=1.77.0
agents>=0.4.0  # OpenAI Agents SDK

# API Framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0

# Database
asyncpg>=0.29.0
psycopg2-binary>=2.9.0  # For migrations
pgvector>=0.2.0  # Python client for pgvector (optional, mainly in-db)

# Messaging/Event Stream
aiokafka>=0.10.0
confluent-kafka>=2.0.0  # Optional alternative

# Channel Integrations
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
twilio>=8.10.0

# Utilities
python-dotenv>=0.19.0
redis>=5.0.0  # Optional: for caching/rate limiting
httpx>=0.27.0  # For external API calls

# Monitoring/Tracing (optional)
prometheus-client>=0.19.0
opentelemetry-api>=1.19.0
opentelemetry-sdk>=1.19.0

# Development/Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
ipdb>=0.13.0

# Security
python-jose[cryptography]>=3.3.0  # JWT tokens
passlib[bcrypt]>=1.7.0  # password hashing

# Data Processing
pandas>=2.0.0  # Optional: for analytics
numpy>=1.24.0  # Numerical ops
